Revert "chore(Core/Spell): SpellId's should always be uint32"
This reverts commit bd19e6bceff6d42eb68f5121df90b21f23692d01.

+++ src/server/game/Server/WorldSession.h
@@ -739,7 +739,7 @@ public:                                                 // opcodes handlers
     //Pet
     void HandlePetAction(WorldPacket& recvData);
     void HandlePetStopAttack(WorldPacket& recvData);
-    void HandlePetActionHelper(Unit* pet, ObjectGuid guid1, uint32 spellid, uint16 flag, ObjectGuid guid2);
+    void HandlePetActionHelper(Unit* pet, ObjectGuid guid1, uint16 spellid, uint16 flag, ObjectGuid guid2);
     void HandlePetNameQuery(WorldPacket& recvData);
     void HandlePetSetAction(WorldPacket& recvData);
     void HandlePetAbandon(WorldPacket& recvData);
